Back/requirements.txt
datetime
pandas
numpy
mysqlclient
python-dotenv
flask
flask-cors


